@@ -1,30 +1,36 @@
 <template>
-<div>
-  <a :href='vm.url' class='url' target='_blank'>{{vm.url}}</a>
-  <div class='iframe-parent' v-html='getDecodedMediaFrame()'></div>
-</div>
+  <div>
+    <a :href="vm.url" class="url" target="_blank">{{ vm.url }}</a>
+    <div class="iframe-parent" v-html="getDecodedMediaFrame()"></div>
+  </div>
 </template>
+
 <script>
-const he = require('he');
+import he from 'he';
 
 export default {
   name: 'MediaViewer',
-  props: ['vm'],
+  props: {
+    vm: {
+      type: Object,
+      required: true
+    }
+  },
   data() {
     return {
       isVisible: false,
       embedded: false,
       decoded: null
-    }
+    };
   },
   methods: {
     getDecodedMediaFrame() {
       if (this.decoded) return this.decoded;
-      this.decoded = he.decode(this.vm.media_embed.content).replace(/position:absolute;/, '')
+      this.decoded = he.decode(this.vm.media_embed.content).replace(/position:absolute;/, '');
       return this.decoded;
     }
   }
-}
+};
 </script>
 
 <style>
