@@ -1,12 +1,13 @@
 <template>
-<video controls class='reddit-video'>
-  <source :src="link" type="video/mp4">
-</video>
+  <video controls class="reddit-video">
+    <source :src="link" type="video/mp4">
+  </video>
 </template>
+
 <script>
 export default {
   name: 'ImgurVideoViewer',
-  props: ['vm'],
+  props: ['vm'], 
   computed: {
     link() {
       return this.vm.url.replace(/\.gifv$/, '.mp4');
@@ -14,5 +15,3 @@ export default {
   },
 }
 </script>
-<style>
-</style>
